reorganized skeletons into a submodule
