+++ meshparty/skeletons/skeletonize.py
@@ -8,430 +8,7 @@ from scipy.spatial import cKDTree as KDTree
 from copy import copy
 import pcst_fast  
 from tqdm import trange, tqdm
-
-def edge_averaged_vertex_property(edge_property, vertices, edges):
-    vertex_property = np.full((len(vertices),2), np.nan)
-    vertex_property[edges[:,0],0] = np.array(edge_property)
-    vertex_property[edges[:,1],1] = np.array(edge_property)
-    return np.nanmean(vertex_property, axis=1)
-    
-def reduce_vertices(vertices, edges, v_filter=None, e_filter=None, return_filter_inds=False):
-    if v_filter is None:
-        v_filter = np.unique(edges).astype(int)
-    if e_filter is None:
-        e_filter_bool = np.isin(edges[:,0], v_filter) & np.isin(edges[:,1], v_filter)
-        e_filter = np.where(e_filter_bool)[0]
-    
-    vertices_n = vertices[v_filter]
-    vmap = dict(zip(v_filter, np.arange(len(v_filter))))
-    
-    edges_f = edges[e_filter].copy()
-    edges_n = np.stack((np.fromiter((vmap[x] for x in edges_f[:,0]), dtype=int),
-                        np.fromiter((vmap[x] for x in edges_f[:,1]), dtype=int))).T
-
-    if return_filter_inds:
-        return vertices_n, edges_n, (v_filter, e_filter)
-    else:
-        return vertice_n, edges_n
-
-def create_csgraph(vertices,
-                   edges,
-                   euclidean_weight=True):
-
-    if euclidean_weight:
-        xs = vertices[edges[:,0]]
-        ys = vertices[edges[:,1]]
-        weights = np.linalg.norm(xs-ys, axis=1)
-        use_dtype = np.float32
-    else:   
-        weights = np.ones((len(edges),)).astype(bool)
-        use_dtype = bool 
-
-    edges = np.concatenate([edges.T, edges.T[[1, 0]]], axis=1)
-    weights = np.concatenate([weights, weights]).astype(dtype=use_dtype)
-
-    csgraph = sparse.csr_matrix((weights, edges),
-                                shape=[len(vertices), ] * 2,
-                                dtype=use_dtype)
-
-    return csgraph
-
-
-class SkeletonForest():
-    def __init__(self, vertices, edges, vertex_properties={}, edge_properties={}, root=None):
-        self._vertex_components = np.full(len(vertices), None)
-        self._edge_components = np.full(len(edges), None)
-        self._skeletons = []
-        self._kdtree = None
-
-        vertices = np.array(vertices)
-        edges = np.array(edges)
-
-        nc, v_lbls = sparse.csgraph.connected_components(create_csgraph(vertices, edges, euclidean_weight=False))
-        lbls, count = np.unique(v_lbls, return_counts=True)
-        lbl_order = np.argsort(count)[::-1]
-        for lbl in lbls[lbl_order]:
-            v_filter = np.where(v_lbls==lbl)[0]
-            vertices_f, edges_f, filters = reduce_vertices(vertices,
-                                                           edges,
-                                                           v_filter=v_filter,
-                                                           return_filter_inds=True)
-            vertex_properties_f = {vp_n: np.array(vp_v)[filters[0]] for vp_n, vp_v in vertex_properties.items()}
-            edge_properties_f = {ep_n: np.array(ep_v)[filters[1]] for ep_n, ep_v in edge_properties.items()}
-
-            self._vertex_components[filters[0]] = lbl
-            self._edge_components[filters[1]] = lbl
-            if root in v_filter:
-                root_f = np.where(root==v_filter)[0][0]
-            else:
-                root_f = None
-            self._skeletons.append( Skeleton(vertices_f, edges_f, vertex_properties_f, edge_properties_f, root=root_f) )
-
-    def __getitem__(self, key):
-        return self._skeletons[key]
-
-    @property
-    def vertices(self):
-        vs_list = [skeleton.vertices for skeleton in self._skeletons]
-        return np.vstack(vs_list)
-
-    @property
-    def edges(self):
-        edges_stacked = []
-        n_shift = 0
-        for skeleton in self._skeletons:
-            edges_stacked.append(skeleton.edges + n_shift)
-            n_shift += skeleton.n_vertices
-        return np.vstack(edges_stacked)
-
-    @property
-    def csgraph(self):    
-        return create_csgraph(self.vertices_all, self.edges_all)
-
-    def vertex_property(self, property_name):
-        vp_list = [skeleton.vertex_properties[property_name] for skeleton in self._skeletons if len(skeleton.vertices)>1]
-        return np.concatenate(vp_list)
-
-    def edge_property(self, property_name):
-        ep_list = [skeleton.edge_properties[property_name] for skeleton in self._skeletons]
-        return ep.concatenate(ep_list)
-
-    @property
-    def kdtree(self):
-        if self._kdtree is None:
-            self._kdtree = KDTree(self.vertices)
-        return self._kdtree
-    
-
-    
-
-class Skeleton:
-    def __init__(self, vertices, edges, vertex_properties={}, edge_properties={}, root=None):
-        self._vertices = np.array(vertices)
-        self._edges = np.vstack(edges).astype(int)
-        self.vertex_properties = vertex_properties
-        self.edge_properties = edge_properties
-
-        self._reset_derived_objects()
-        if root is None:
-            self._create_default_root()
-        else:
-            self.reroot(root)
-
-        # self._parent_node = {}
-        self._kdtree = None
-        self._branch_points = None
-        self._end_points = None
-
-    @property
-    def vertices(self):
-        return self._vertices.copy()
-
-    @property
-    def edges(self):
-        return self._edges.copy()
-
-    @property
-    def csgraph(self):
-        if self._csgraph is None:
-            self._csgraph = self._create_csgraph()
-        return self._csgraph.copy()
-    
-    @property
-    def csgraph_binary(self):
-        if self._csgraph_binary is None:
-            self._csgraph_binary = self._create_csgraph(euclidean_weight=False)
-        return self._csgraph_binary
-
-    @property
-    def csgraph_undirected(self):
-        return self.csgraph + self.csgraph.T
-
-    @property
-    def csgraph_binary_undirected(self):
-        return self.csgraph_binary + self.csgraph_binary.T
-
-    @property
-    def n_vertices(self):
-        return len(self.vertices)
-
-    @property
-    def root(self):
-        if self._root is None:
-            self._create_default_root()
-        return copy(self._root)
-
-    @property
-    def kdtree(self):
-        if self._kdtree is None:
-            self._kdtree = KDTree(self.vertices)
-        return self._kdtree 
-
-    def _create_default_root(self):
-        r = find_far_points_graph(self._create_csgraph(directed=False))
-        self.reroot(r[0])
-
-    def _parent_node(self, vind):
-        return self._parent_node_array[vind]
-
-    def reroot(self, new_root):
-        self._root = new_root
-        self._parent_node_array = np.full(self.n_vertices, None)
-        
-        # The edge list has to be treated like an undirected graph
-        d = sparse.csgraph.dijkstra(self.csgraph_binary,
-                                    directed=False,
-                                    indices=new_root)
-
-        # Make edges in edge list orient as [child, parent]
-        # Where each child only has one parent
-        # And the root has no parent. (Thus parent is closer than child)
-        edges = self.edges
-        is_ordered = d[edges[:,0]] > d[edges[:,1]]
-        e1 = np.where( is_ordered, edges[:,0], edges[:,1])
-        e2 = np.where( is_ordered, edges[:,1], edges[:,0])
-
-        self._edges = np.stack((e1,e2)).T
-        self._parent_node_array[e1]=e2
-        self._reset_derived_objects()
-
-
-    def _reset_derived_objects(self):
-        self._csgraph = None
-        self._csgraph_binary = None
-        self._paths = None
-
-    def _create_csgraph(self,
-                        directed=True,
-                        euclidean_weight=True):
-
-        edges = self._edges.copy()
-
-        if euclidean_weight:
-            xs = self.vertices[edges[:,0]]
-            ys = self.vertices[edges[:,1]]
-            weights = np.linalg.norm(xs-ys, axis=1)
-            use_dtype = np.float32
-        else:   
-            weights = np.ones((len(edges),)).astype(bool)
-            use_dtype = bool 
-
-        if directed:
-            edges = edges.T
-        else:
-            edges = np.concatenate([edges.T, edges.T[[1, 0]]], axis=1)
-            weights = np.concatenate([weights, weights]).astype(dtype=use_dtype)
-
-        csgraph = sparse.csr_matrix((weights, edges),
-                                    shape=[len(self.vertices), ] * 2,
-                                    dtype=use_dtype)
-
-        return csgraph
-
-
-    @property
-    def branch_points(self):
-        if self._branch_points is None:
-            self._create_branch_and_end_points()
-        return self._branch_points.copy()
-
-
-    @property
-    def n_branch_points(self):
-        if self._branch_points is None:
-            self._create_branch_and_end_points()
-        return len(self._branch_points)
-
-    @property
-    def end_points(self):
-        if self._end_points is None:
-            self._create_branch_and_end_points()
-        return self._end_points.copy()
-
-    @property
-    def n_end_points(self):
-        if self._end_points is None:
-            self._create_branch_and_end_points()
-        return len(self._end_points)
-
-    def _create_branch_and_end_points(self):
-        n_children = np.sum(self.csgraph_binary>0, axis=0).squeeze()
-        self._branch_points = np.flatnonzero(n_children > 1)
-        self._end_points = np.flatnonzero(n_children == 0)
-
-    @property
-    def paths(self):
-        if self._paths is None:
-            self._paths = self._compute_paths()
-        return self._paths.copy()
-    
-    def distance_to_root(self, indices):
-        ds = sparse.csgraph.dijkstra(self.csgraph, directed=True, indices=indices)
-        return ds[:,self.root]
-
-    def path_to_root(self, v_ind):
-        '''
-        Returns an ordered path to root from a given vertex node.
-        '''
-        path = [v_ind]
-        ind = v_ind
-        while ind is not None:
-            ind = self._parent_node(ind)
-            path.append(ind)
-        return path
-
-    def path_length(self, paths=None):
-        if paths is None:
-            paths = self.paths
-        L = 0
-        for path in paths:
-            L += self._single_path_length(path)
-        return L
-
-    def _single_path_length(self, path):
-        xs = self.vertices[ path[:-1] ]
-        ys = self.vertices[ path[1:] ]
-        return sum(np.linalg.norm(ys-xs))
-
-    def _compute_paths(self):
-        '''
-        Only considers the component with root
-        '''
-        ds, P = sparse.csgraph.dijkstra(self.csgraph,
-                                        directed=True,
-                                        indices=self.end_points,
-                                        return_predecessors=True)
-        d_to_root = ds[:,self.root]
-        end_point_order = np.argsort(d_to_root)[::-1]
-        paths = []
-
-        visited = np.full(shape=(len(self.vertices),), fill_value=False)
-        visited[self.root] = True
-        for ep_ind in end_point_order:
-            if np.isinf(d_to_root[ep_ind]):
-                continue
-            path, visited = self._unvisited_path_on_tree(self.end_points[ep_ind],
-                                                         visited)
-            paths.append(path)
-
-        return paths
-
-    def _unvisited_path_on_tree(self, ind, visited):
-        '''
-            Find path from ind to a visited node along G
-            Assumes that G[i,j] means i->j
-        '''
-        G = self.csgraph
-        n_ind = ind
-        path = [n_ind]
-        while visited[n_ind] == False:
-            visited[n_ind] = True
-            n_ind = self._parent_node(n_ind)
-            path.append(n_ind)
-        return path, visited
-
-
-def connected_component_slice(G, ind=None):
-    _, labels = sparse.csgraph.connected_components(G)
-    if ind is None:
-        label_vals, cnt = np.unique(labels, return_counts=True)
-        ind = np.argmax(cnt)
-        label = label_vals[ind]
-    else:
-        label = labels[ind]
-    return np.where(labels == label)[0]
-
-
-def get_path(root, target, pred):
-    path = [target]
-    p = target
-    while p != root:
-        p = pred[p]
-        path.append(p)
-    path.reverse()
-    return path
-
-
-def find_far_points(trimesh, multicomponent=False):
-    return find_far_points_graph(trimesh.csgraph)
-
-
-def find_far_points_graph(mesh_graph, multicomponent=False):
-    d = 0
-    dn = 1
-   
-    if multicomponent: 
-        a = connected_component_slice(mesh_graph)[0]
-    else:
-        a = 0
-    b = 1
-    
-    k = 0
-    pred = None
-    ds = None
-    while 1:
-        k += 1
-        dsn, predn = sparse.csgraph.dijkstra(
-            mesh_graph, False, a, return_predecessors=True)
-        if multicomponent:
-            dsn[np.isinf(dsn)] = 0
-        bn = np.argmax(dsn)
-        dn = dsn[bn]
-        if dn > d:
-            b = a
-            a = bn
-            d = dn
-            pred = predn
-            ds = dsn
-        else:
-            break
-
-    return b, a, pred, d, ds
-
-
-
-def setup_root(mesh, soma_pos=None, soma_thresh=7500):
-    valid = np.ones(len(mesh.vertices), np.bool)
-
-    # soma mode
-    if soma_pos is not None:
-        # pick the first soma as root
-        soma_d, soma_i = mesh.kdtree.query(soma_pos,
-                                           k=len(mesh.vertices),
-                                           distance_upper_bound=soma_thresh)
-        assert(np.min(soma_d)<soma_thresh)
-        root = soma_i[np.argmin(soma_d)]
-        valid[soma_i[~np.isinf(soma_d)]] = False
-        root_ds, pred = sparse.csgraph.dijkstra(mesh.csgraph,
-                                                False,
-                                                root,
-                                                return_predecessors=True)
-    else:
-        # there is no soma so use far point heuristic
-        root, target, pred, dm, root_ds = find_far_points(mesh)
-        valid[root] = 0
-
-    return root, root_ds, pred, valid
+from meshparty.skeletons import sk_utils
 
 
 def recenter_verts(verts, edges, centers):
@@ -852,15 +429,6 @@ def mesh_teasar(mesh, root=None, valid=None, root_ds=None, root_pred=None, soma_
         return paths, path_lengths, time_arrays, dt
     else:
         return paths, path_lengths
-
-
-def paths_to_edges(path_list):
-    arrays = []
-    for path in path_list:
-        p = np.array(path)
-        e = np.vstack((p[0:-1], p[1:])).T
-        arrays.append(e)
-    return np.vstack(arrays)
 
 
 def smooth_graph(verts, edges, neighborhood=2, iterations=100, r=.1):
